抽出聊天路由
