+++ api-service/src/chat/mod.rs
@@ -1,1 +1,2 @@
+pub mod control;
 pub mod message;
